refactor(desktop): drain project catalog reconciliation
Track explicit reconcile completion through catalog shutdown so scans cannot outlive their catalog or test-owned workspace. Split manual ordering and registration logic into focused files to keep repository structure budgets unchanged.

Verify shutdown drainage and the formerly flaky pinned-session-owner cleanup path across repeated runs.
