+++ internal/sessioncatalog/catalog.go
@@ -503,156 +503,6 @@ func mapKeys(values map[string]struct{}) []string {
 	}
 	return out
 }
-func (c *Catalog) ListTopics(ctx context.Context, req TopicPageRequest) (TopicPage, error) {
-	out := TopicPage{Items: []TopicRecord{}, Revision: c.revision.Load()}
-	req.Scope, req.WorkspaceRoot = normalizeScope(req.Scope, req.WorkspaceRoot)
-	if req.Limit <= 0 {
-		req.Limit = DefaultLimit
-	}
-	if req.Limit > MaxLimit {
-		req.Limit = MaxLimit
-	}
-	cursor, err := decodeCursor(req.Cursor)
-	if err != nil {
-		return out, err
-	}
-	if cursor != nil && cursor.ManualOrder != req.ManualOrder {
-		return out, errCursorSortModeChanged
-	}
-	args := []any{req.Scope, req.WorkspaceRoot}
-	where := `scope=? AND workspace_root=?`
-	if query := strings.TrimSpace(req.Query); query != "" {
-		where += ` AND lower(title) LIKE ?`
-		args = append(args, "%"+strings.ToLower(query)+"%")
-	}
-	if cutoff := timeFilterCutoff(req.TimeFilter, c.opts.Now()); cutoff > 0 {
-		where += ` AND last_activity_at>=?`
-		args = append(args, cutoff)
-	}
-	sortExpression := topicPageSortExpression(req.SortMode)
-	manualSortExpression := topicPageManualSortExpression()
-	scanCursor := cursor
-	scanLimit := max(req.Limit+1, 64)
-	for len(out.Items) <= req.Limit {
-		pageWhere := where
-		pageArgs := append([]any(nil), args...)
-		if scanCursor != nil {
-			if req.ManualOrder {
-				pageWhere += ` AND (pinned<? OR (pinned=? AND ` + manualSortExpression + `>?) OR ` +
-					`(pinned=? AND ` + manualSortExpression + `=? AND ` + sortExpression + `<?) OR ` +
-					`(pinned=? AND ` + manualSortExpression + `=? AND ` + sortExpression + `=? AND topic_id>?))`
-				pageArgs = append(pageArgs,
-					scanCursor.Pinned,
-					scanCursor.Pinned, scanCursor.SortOrder,
-					scanCursor.Pinned, scanCursor.SortOrder, scanCursor.Activity,
-					scanCursor.Pinned, scanCursor.SortOrder, scanCursor.Activity, scanCursor.TopicID,
-				)
-			} else {
-				pageWhere += ` AND (pinned<? OR (pinned=? AND ` + sortExpression + `<?) OR (pinned=? AND ` + sortExpression + `=? AND topic_id>?))`
-				pageArgs = append(pageArgs, scanCursor.Pinned, scanCursor.Pinned, scanCursor.Activity,
-					scanCursor.Pinned, scanCursor.Activity, scanCursor.TopicID)
-			}
-		}
-		pageArgs = append(pageArgs, scanLimit)
-		orderBy := `pinned DESC,` + sortExpression + ` DESC,topic_id ASC`
-		if req.ManualOrder {
-			orderBy = `pinned DESC,` + manualSortExpression + ` ASC,` + sortExpression + ` DESC,topic_id ASC`
-		}
-		rows, err := c.db.QueryContext(ctx, `SELECT scope,workspace_root,topic_id,title,title_source,pinned,
-			CASE WHEN metadata_present=1 THEN sort_order ELSE -1 END,
-            turns,turns_state,created_at,last_activity_at,recovery_state,recovery_branch_count,
-            recovery_unresolved_count,recovery_cleanup_eligible_count,health
-            FROM catalog_topics WHERE `+pageWhere+`
-			ORDER BY `+orderBy+` LIMIT ?`, pageArgs...)
-		if err != nil {
-			return out, err
-		}
-		// Drain the page before hydrating sessions: an open cursor holds a
-		// connection, and listTopicSessions needs a second one, so nesting them
-		// deadlocks whenever the pool is saturated (always in memory mode).
-		scanned := make([]TopicRecord, 0, scanLimit)
-		for rows.Next() {
-			var item TopicRecord
-			if err := rows.Scan(&item.Scope, &item.WorkspaceRoot, &item.TopicID, &item.Title,
-				&item.TitleSource, &item.Pinned, &item.SortOrder, &item.Turns, &item.TurnsState,
-				&item.CreatedAt, &item.LastActivityAt, &item.RecoveryState, &item.RecoveryBranchCount,
-				&item.RecoveryUnresolvedCount, &item.RecoveryCleanupEligibleCount, &item.Health); err != nil {
-				_ = rows.Close()
-				return out, err
-			}
-			scanned = append(scanned, item)
-		}
-		rowsErr := rows.Err()
-		_ = rows.Close()
-		if rowsErr != nil {
-			return out, rowsErr
-		}
-		rawCount := len(scanned)
-		overflow := false
-		for _, item := range scanned {
-			sessions, err := c.listTopicSessions(ctx, TopicKey{
-				Scope: item.Scope, WorkspaceRoot: item.WorkspaceRoot, TopicID: item.TopicID,
-			})
-			if err != nil {
-				return TopicPage{Items: []TopicRecord{}, Revision: out.Revision}, err
-			}
-			if len(sessions) == 0 {
-				continue
-			}
-			// Skip pure recovery-shell topics that lost their ordinary
-			// representative after lineage re-anchoring. Prevents empty pages
-			// and flash of conflict rows while catalog rebuilds.
-			hasOrdinary := false
-			for _, session := range sessions {
-				if session.OrdinaryVisible || (!session.Recovered && !session.RecoveryCopy) {
-					hasOrdinary = true
-					break
-				}
-			}
-			if !hasOrdinary && !item.Pinned {
-				continue
-			}
-			item.Sessions = sessions
-			hydrateTopicDisplay(&item)
-			out.Items = append(out.Items, item)
-			if len(out.Items) > req.Limit {
-				overflow = true
-				break
-			}
-		}
-		if overflow || rawCount < scanLimit || rawCount == 0 {
-			break
-		}
-		lastScanned := scanned[rawCount-1]
-		pinned := 0
-		if lastScanned.Pinned {
-			pinned = 1
-		}
-		scanCursor = &pageCursor{
-			Pinned: pinned, ManualOrder: req.ManualOrder,
-			SortOrder: topicPageManualSortValue(lastScanned),
-			Activity:  topicPageSortValue(lastScanned, req.SortMode), TopicID: lastScanned.TopicID,
-		}
-	}
-	more := len(out.Items) > req.Limit
-	if more {
-		out.Items = out.Items[:req.Limit]
-	}
-	if more && len(out.Items) > 0 {
-		last := out.Items[len(out.Items)-1]
-		pinned := 0
-		if last.Pinned {
-			pinned = 1
-		}
-		out.NextCursor = encodeCursor(pageCursor{
-			Pinned: pinned, ManualOrder: req.ManualOrder,
-			SortOrder: topicPageManualSortValue(last),
-			Activity:  topicPageSortValue(last, req.SortMode), TopicID: last.TopicID,
-		})
-	}
-	return out, nil
-}
-
 func (c *Catalog) listTopicSessions(ctx context.Context, key TopicKey) ([]SessionRecord, error) {
 	out := []SessionRecord{}
 	var cursor *sessionPageCursor
